Render SVG images too
The SVG images are compressed but do not indicate that by file extension,
so mime type guessing by file extension is no good, so do_GET() from
SimpleHTTPServer is not enough.  The SVG images also contain some Adobe
Illustrator specific gunk which at least Firefox refuses to handle at all,
so we need to massage that a bit while serving it.

Luckily all other content seems to be nicely typed so we can just
assume octet-stream to mean compressed SVG. Use gzip module to read,
and parse with entity resolving off to filter out the bad apples,
and feed the result to the now happy browser.

Refactor do_page() to handle different content types and take advantage
of that in all serving of our own.