Render options (aka codep in elearn jargon) in titles/links, more or less
Optional features like alarm, sw/sedan, xenon lights etc are handled
with "CODEP" in eLearn: where ALL_CODEP=0 there's a related CODEP
entry to be found via ELEMENT_CODEP relation. Fish the codeps
in content menus and for the path view too, show in parentheses
after the main title.

This seems awfully clumsy, but I dont see a way to perform this
magic in a nice and elegant sql query (maybe there is one, maybe
there is not) so we patch it in as needed. To make matters uglier,
sqlite.Row does not support assignment so we'll convert to dicts
where codep handling is needed.