Regenerate searchable text in XML table
The FULL_TEXT column in XML table is supposed to be type VARCHAR(255)
according to the schema, but in reality it is an arbitrary long
field just like MEMO type, but mdbtools doesn't know how to process
the data beyond the predefined limit. As a result the FULL_TEXT
contents are truncated to a pitiful 255 characters which makes it
practically useless for its original purpose: searching.

So we add another helper script which performs the XSL transformations
on all the XML entries, extracts the plain text from the resulting
html and updates the table.